Separate pysmurf_controller and pysmurf_monitor
