+++ agents/pysmurf_monitor/requirements.txt
@@ -1,0 +1,1 @@
+mysql-connector>=2.1.6
